Handle Vessels Full Cases

+++ Models/Dtos/ShipmentRequest/CreateShipmentRequestDto.cs
@@ -2,7 +2,7 @@ namespace RiverLine.Api.Models.Dtos.ShipmentRequest;
 
 public record CreateShipmentRequestDto(
     string CargoType,
-    double Weight,
+    decimal Weight,
     Guid OriginNileBerthId,
     Guid DestinationNileBerthId,
     DateOnly RequestedDate
